+++ liboath/tests/tst_usersfile.c
@@ -136,8 +136,8 @@ main (void)
     }
   if (last_otp != 1260206742)
     {
-      printf ("oath_authenticate_usersfile timestamp %ld != 1260203142\n",
-	      last_otp);
+      printf ("oath_authenticate_usersfile timestamp %lu != 1260203142\n",
+	      (unsigned long) last_otp);
       return 1;
     }
